# ARROW-7491: [Java] Improve the performance of aligning