Aligning is an important and frequent operation when writing IPC data. It writes no more than 7 0 bytes to the output.
The current implementation creates a new byte array each time, leading to performance overhead, and increases the GC pressure.

We improve it by means of a shared byte array. Benchmark evaluation shows a 11% performance gain.