@liyafan82 2020-01-03
Performance data:

before:
Benchmark                             (alignSize)  Mode  Cnt    Score   Error  Units
WriteChannelBenchmark.alignBenchmark            1  avgt    5  139.916 ± 0.721  ns/op
WriteChannelBenchmark.alignBenchmark            2  avgt    5  132.845 ± 1.009  ns/op
WriteChannelBenchmark.alignBenchmark            3  avgt    5  147.814 ± 0.559  ns/op
WriteChannelBenchmark.alignBenchmark            4  avgt    5  134.697 ± 0.852  ns/op
WriteChannelBenchmark.alignBenchmark            5  avgt    5  145.638 ± 1.760  ns/op
WriteChannelBenchmark.alignBenchmark            6  avgt    5  159.414 ± 1.750  ns/op
WriteChannelBenchmark.alignBenchmark            7  avgt    5  147.423 ± 1.451  ns/op

after:
Benchmark                             (alignSize)  Mode  Cnt    Score   Error  Units
WriteChannelBenchmark.alignBenchmark            1  avgt    5  122.111 ± 1.474  ns/op
WriteChannelBenchmark.alignBenchmark            2  avgt    5  123.116 ± 0.970  ns/op
WriteChannelBenchmark.alignBenchmark            3  avgt    5  123.983 ± 1.485  ns/op
WriteChannelBenchmark.alignBenchmark            4  avgt    5  124.089 ± 1.165  ns/op
WriteChannelBenchmark.alignBenchmark            5  avgt    5  129.945 ± 0.436  ns/op
WriteChannelBenchmark.alignBenchmark            6  avgt    5  132.752 ± 1.037  ns/op
WriteChannelBenchmark.alignBenchmark            7  avgt    5  141.913 ± 0.755  ns/op