@fsaintjacques 2020-01-06
This change is pointless if `align` is not called often in an IPC stream. OTOH, you found a legitimate issue in writeZeroes, I'd fix the `writeZeros` function with a static constant buffer (like you did) and a loop instead. Otherwise one might have surprise if it's called with say `writeZeroes(2^31)`.