@liyafan82 2020-01-07
> This change is pointless if `align` is not called often in an IPC stream. OTOH, you found a legitimate issue in writeZeroes, I'd fix the `writeZeros` function with a static constant buffer (like you did) and a loop instead. Otherwise one might have surprise if it's called with say `writeZeroes(2^31)`.

@fsaintjacques Thanks a lot for you feedback.

The number of calls to `align` depends on the distribution of the size for each write operation. If the write size is uniformly distributed, there is a 7/8 probability that the `align` method will be called. 

In practice, I found the data buffer size is often rounded to a multiple of 8, but the validity buffer is often not (Maybe this problem will be changed in the future).

I am not sure how often is the `writeZeros` method called. So if you think it is important, maybe you can open an issue to track it. And please note that we can only use a constant array, not a constant buffer, because a buffer can only be consumed once (We can reuse the buffer after calling the `rewind` method, but this will lead to problems in multi-threading scenarios).