@fsaintjacques 2020-01-07
What I'm suggesting is that you preserve the current implementation of align to call writeZeros and fix writeZeros to not allocate a byte array.

Not because this will improve `align` speed, but because writeZeros is O(n) memory when it could be O(1).