@emkornfield 2020-02-03
@fsaintjacques were you going to finish the review here?  Maybe @andygrove can help?