@fsaintjacques 2020-02-03
I'm neutral on this change, the code path still always allocate a temporary ByteBuffer object on the heap via ByteBuffer.wrap.