@liyafan82 2020-02-20
> I'm neutral on this change, the code path still always allocate a temporary ByteBuffer object on the heap via ByteBuffer.wrap.

@fsaintjacques Thanks a lot for your feedback. 
You are right that a temporary ByteBuffer is created for each align operation.

This could have been overcome by creating 7 ByteBuffer objects for each possible size of aligning.

Our consideration is that, creating the ByteBuffer objects can be a waste of resources, as not all of them will be used, and this increases the cost of initializing the WriteChannel object. So reusing the zero byte array should be a good trade-off point. 

Could you please give your opinion about this?