@fsaintjacques 2020-02-20
I spent a lot of time looking back and forth on this patch for the very small gain it'll bring (I think).